refactor: extract Strength tab to views/StrengthView.jsx + test (PR8)
Week-3 burndown (#52). Moves the `strength` tab — lift selector, e1RM trend
chart, weekly strength template, and the logged-session list — out of the
erg-dashboard.jsx monolith into a self-contained component.

- web/src/constants/strength.js: NEW — the strengthTrend (per-lift e1RM seed
  data) and LIFT_COLOR map, lifted out of the monolith into the constants
  layer (both are pure data). strengthTrend is shared with the Overview tab,
  so it lives in constants/ rather than being prop-drilled to two views.
- web/src/views/StrengthView.jsx: the strength JSX, verbatim. Takes the lifted
  state (activeLift/setActiveLift) and the filtered strengthSessions as props;
  derives liftColor internally; imports its data + STRENGTH_TEMPLATES/PREHAB_NOTE
  + LogEntry + StrengthTooltip + Recharts directly.
- web/src/erg-dashboard.jsx: the ~270-line block becomes
  `{view === 'strength' && <StrengthView ... />}`; the monolith now imports
  strengthTrend from constants/strength.js for the Overview readout, and the
  now-unused PREHAB_NOTE/STRENGTH_TEMPLATES/LIFT_COLOR imports are pruned.

Adds web/src/views/__tests__/StrengthView.test.jsx (5 tests): a selector
button per lift, the active-lift e1RM header, setActiveLift firing on select,
the session log rendering, and the colour-map guard. Coverage: 91.7% lines /
83.3% branch / 87.5% funcs.

Build, full Vitest suite (295 passing), lint, and format:check all green.

Closes #75
