+++ web/src/constants/strength.js
@@ -1,0 +1,51 @@
+// ── STRENGTH SEED DATA ─────────────────────────────────────────
+// e1RM (estimated 1-rep-max) progression per lift, and the per-lift
+// accent colours used by the Strength view's trend chart + selector.
+export const strengthTrend = {
+  'Back Squat': [
+    { date: '6/3', e1rm: 109.7 },
+    { date: '6/5', e1rm: 110.5 },
+    { date: '6/9', e1rm: 118.0 },
+  ],
+  'Romanian Deadlift': [
+    { date: '6/3', e1rm: 85.2 },
+    { date: '6/5', e1rm: 89.2 },
+    { date: '6/9', e1rm: 95.8 },
+  ],
+  'Bench Press': [
+    { date: '6/3', e1rm: 63.4 },
+    { date: '6/8', e1rm: 71.2 },
+  ],
+  'Incline Bench': [
+    { date: '6/3', e1rm: 52.9 },
+    { date: '6/7', e1rm: 55.8 },
+    { date: '6/8', e1rm: 61.7 },
+  ],
+  'Cable Row': [
+    { date: '5/31', e1rm: 62.3 },
+    { date: '6/4', e1rm: 74.6 },
+    { date: '6/8', e1rm: 98.0 },
+  ],
+  'Barbell Row': [
+    { date: '5/31', e1rm: 53.8 },
+    { date: '6/4', e1rm: 62.8 },
+  ],
+  'Lat Pulldown': [
+    { date: '5/31', e1rm: 66.1 },
+    { date: '6/4', e1rm: 74.6 },
+  ],
+  'Shoulder Press': [
+    { date: '6/4', e1rm: 41.9 },
+    { date: '6/7', e1rm: 46.2 },
+  ],
+};
+
+export const LIFT_COLOR = {
+  'Back Squat': '#34d399',
+  'Romanian Deadlift': '#34d399',
+  'Bench Press': '#a78bfa',
+  'Incline Bench': '#a78bfa',
+  'Barbell Row': '#00d4ff',
+  'Lat Pulldown': '#00d4ff',
+  'Shoulder Press': '#f472b6',
+};
